requirements.txt
gradio
matplotlib
pandas
numpy
google-genai
python-dotenv